Moved Pie Labels
Increased the length of the grey pie chart label "ticks", and increased the spacing between the percentages and labels, as well as between the ticks and percentages or labels.  Added "Total" so that the center now reads "Total \n Responses: \n <#>".  Changed ValusFS & TextFS to 18px (used to be 20px).  This makes the labels a little smaller, but work with the y-attribute spacing I've calculated and entered in.